+++ src/client/CMakeLists.txt
@@ -1,5 +1,12 @@
 set(client_SRCS
-	${CMAKE_CURRENT_SOURCE_DIR}/renderingcore.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/render/anaglyph.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/render/core.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/render/factory.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/render/interlaced.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/render/pageflip.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/render/plain.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/render/sidebyside.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/render/stereo.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/renderingengine.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/clientlauncher.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/inputhandler.cpp
@@ -7,4 +14,3 @@ set(client_SRCS
 	${CMAKE_CURRENT_SOURCE_DIR}/joystick_controller.cpp
 	PARENT_SCOPE
 )
-
