+++ src/client/render/core.cpp
@@ -1,0 +1,105 @@
+/*
+Minetest
+Copyright (C) 2010-2013 celeron55, Perttu Ahola <celeron55@gmail.com>
+Copyright (C) 2017 numzero, Lobachesky Vitaly <numzer0@yandex.ru>
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation; either version 2.1 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+
+#include "core.h"
+#include "camera.h"
+#include "client.h"
+#include "clientmap.h"
+#include "hud.h"
+#include "minimap.h"
+
+RenderingCore::RenderingCore(IrrlichtDevice *_device) :
+	device(_device),
+	driver(device->getVideoDriver()),
+	smgr(device->getSceneManager()),
+	guienv(device->getGUIEnvironment())
+{
+	screensize = driver->getScreenSize();
+}
+
+RenderingCore::~RenderingCore()
+{
+	clear_textures();
+}
+
+void RenderingCore::initialize(Client *_client, Hud *_hud)
+{
+	client = _client;
+	camera = client->getCamera();
+	mapper = client->getMinimap();
+	hud = _hud;
+// have to be called now as the VMT is not ready in the constructor:
+	init_textures();
+}
+
+void RenderingCore::update_screen_size()
+{
+	clear_textures();
+	init_textures();
+}
+
+void RenderingCore::draw(video::SColor _skycolor, bool _show_hud, bool _show_minimap,
+		bool _draw_wield_tool, bool _draw_crosshair)
+{
+	v2u32 ss = driver->getScreenSize();
+	if (screensize != ss) {
+		screensize = ss;
+		update_screen_size();
+	}
+	skycolor = _skycolor;
+	show_hud = _show_hud;
+	show_minimap = _show_minimap;
+	draw_wield_tool = _draw_wield_tool;
+	draw_crosshair = _draw_crosshair;
+
+	pre_draw();
+	draw_all();
+	post_draw();
+}
+
+void RenderingCore::draw_3d()
+{
+	smgr->drawAll();
+	driver->setTransform(video::ETS_WORLD, core::IdentityMatrix);
+	if (!show_hud)
+		return;
+	hud->drawSelectionMesh();
+	if (draw_wield_tool)
+		camera->drawWieldedTool();
+}
+
+void RenderingCore::draw_hud()
+{
+	if (!show_hud)
+		return;
+	if (draw_crosshair)
+		hud->drawCrosshair();
+	hud->drawHotbar(client->getPlayerItem());
+	hud->drawLuaElements(camera->getOffset());
+	camera->drawNametags();
+	if (mapper && show_minimap)
+		mapper->drawMinimap();
+	guienv->drawAll();
+}
+
+void RenderingCore::draw_last_fx()
+{
+	client->getEnv().getClientMap().renderPostFx(camera->getCameraMode());
+}
