+++ src/client/render/core.h
@@ -71,91 +71,3 @@ public:
 	void draw(video::SColor _skycolor, bool _show_hud, bool _show_minimap,
 			bool _draw_wield_tool, bool _draw_crosshair);
 };
-
-class RenderingCorePlain: public RenderingCore
-{
-public:
-	using RenderingCore::RenderingCore;
-	void draw_all() override;
-};
-
-class RenderingCoreStereo: public RenderingCore
-{
-protected:
-	scene::ICameraSceneNode *cam;
-	core::matrix4 base_transform;
-	float parallax_strength;
-
-	void pre_draw();
-	virtual void use_eye(bool right);
-	virtual void reset_eye();
-	void render_two();
-
-public:
-	RenderingCoreStereo(IrrlichtDevice *_device);
-};
-
-class RenderingCoreAnaglyph: public RenderingCoreStereo
-{
-protected:
-	void setup_material(int color_mask);
-	void use_eye(bool right) override;
-	void reset_eye() override;
-
-public:
-	using RenderingCoreStereo::RenderingCoreStereo;
-	void draw_all() override;
-};
-
-class RenderingCoreSideBySide: public RenderingCoreStereo
-{
-protected:
-	video::ITexture *left = nullptr;
-	video::ITexture *right = nullptr;
-	video::ITexture *hud = nullptr;
-
-	void init_textures() override;
-	void clear_textures() override;
-	void use_eye(bool right) override;
-	void reset_eye() override;
-
-public:
-	using RenderingCoreStereo::RenderingCoreStereo;
-	void draw_all() override;
-};
-
-class RenderingCorePageflip: public RenderingCoreStereo
-{
-protected:
-	video::ITexture *hud = nullptr;
-
-	void init_textures() override;
-	void clear_textures() override;
-	void use_eye(bool right) override;
-	void reset_eye() override;
-
-public:
-	using RenderingCoreStereo::RenderingCoreStereo;
-	void draw_all() override;
-};
-
-class RenderingCoreInterlaced: public RenderingCoreStereo
-{
-protected:
-	video::ITexture *left = nullptr;
-	video::ITexture *right = nullptr;
-	video::ITexture *mask = nullptr;
-	video::SMaterial mat;
-
-	void init_material();
-	void init_textures() override;
-	void clear_textures() override;
-	void init_mask();
-	void use_eye(bool right) override;
-	void reset_eye() override;
-	void merge();
-
-public:
-	RenderingCoreInterlaced(IrrlichtDevice *_device);
-	void draw_all() override;
-};
