+++ src/client/render/interlaced.h
@@ -1,0 +1,43 @@
+/*
+Minetest
+Copyright (C) 2010-2013 celeron55, Perttu Ahola <celeron55@gmail.com>
+Copyright (C) 2017 numzero, Lobachesky Vitaly <numzer0@yandex.ru>
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation; either version 2.1 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+
+#pragma once
+#include "stereo.h"
+
+class RenderingCoreInterlaced: public RenderingCoreStereo
+{
+protected:
+	video::ITexture *left = nullptr;
+	video::ITexture *right = nullptr;
+	video::ITexture *mask = nullptr;
+	video::SMaterial mat;
+
+	void init_material();
+	void init_textures() override;
+	void clear_textures() override;
+	void init_mask();
+	void use_eye(bool right) override;
+	void reset_eye() override;
+	void merge();
+
+public:
+	RenderingCoreInterlaced(IrrlichtDevice *_device);
+	void draw_all() override;
+};
